# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NFT Gift Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  telegramId: string;
  /** @nullable */
  username?: string | null;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  isNewUser: boolean;
  onboardingComplete: boolean;
  isAdmin?: boolean;
  createdAt: string;
}

export interface UserRegister {
  telegramId: string;
  /** @nullable */
  username?: string | null;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  initData?: string | null;
}

export interface OnboardingUpdate {
  telegramId: string;
}

export interface Gift {
  id: number;
  name: string;
  giftSlug: string;
  telegramLink?: string;
  imageUrl: string;
  /** @nullable */
  attributes?: string | null;
  price: number;
  isListed: boolean;
  createdAt: string;
}

export interface GiftInput {
  name: string;
  giftSlug: string;
  /** @nullable */
  telegramLink?: string | null;
  imageUrl: string;
  /** @nullable */
  attributes?: string | null;
  price: number;
  isListed?: boolean;
}

export interface GiftUpdate {
  price?: number;
  isListed?: boolean;
  /** @nullable */
  name?: string | null;
}

export interface GiftPreviewRequest {
  link: string;
}

export interface GiftPreview {
  name: string;
  giftSlug: string;
  /** @nullable */
  giftNumber?: number | null;
  telegramLink?: string;
  imageUrl: string;
  /** @nullable */
  bgColor?: string | null;
  /** @nullable */
  attributes?: string | null;
}

export interface Promo {
  id: number;
  code: string;
  discountPercent: number;
  maxUses: number;
  usedCount: number;
  /** @nullable */
  expiresAt?: string | null;
  createdBy: string;
  createdAt: string;
}

export interface PromoInput {
  code: string;
  discountPercent: number;
  maxUses?: number;
  /** @nullable */
  expiresAt?: string | null;
}

export interface PromoApplyRequest {
  code: string;
}

export interface PromoApplyResponse {
  valid: boolean;
  discountPercent: number;
  message: string;
}

export interface MarketStats {
  totalGifts: number;
  totalListedGifts: number;
  totalUsers: number;
  /** @nullable */
  minPrice?: number | null;
  /** @nullable */
  maxPrice?: number | null;
}

export type ListGiftsParams = {
  search?: string;
  sortBy?: string;
  sortOrder?: string;
};