# lib/api-zod/src/generated/types/giftPreview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NFT Gift Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface GiftPreview {
  name: string;
  giftSlug: string;
  /** @nullable */
  giftNumber?: number | null;
  telegramLink?: string;
  imageUrl: string;
  /** @nullable */
  bgColor?: string | null;
  /** @nullable */
  attributes?: string | null;
}