# lib/api-zod/src/generated/types/promo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NFT Gift Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Promo {
  id: number;
  code: string;
  discountPercent: number;
  maxUses: number;
  usedCount: number;
  /** @nullable */
  expiresAt?: string | null;
  createdBy: string;
  createdAt: string;
}