# lib/api-zod/src/generated/types/promoInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NFT Gift Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface PromoInput {
  code: string;
  discountPercent: number;
  maxUses?: number;
  /** @nullable */
  expiresAt?: string | null;
}